+++ ChangeLog.md
@@ -8,6 +8,7 @@
     * Loosen import versions of dependencies.
   * Correctly adhere to the Multi-release JAR spec by ensuring all public classes under versioned directories preside over classes present in the top-level directory.
   * Eliminate stray `System.err.println()` calls.
+  * Change PageantConnector to use JNA's built-in support for `User32.SendMessage()`.
 * [0.2.8](https://github.com/mwiede/jsch/releases/tag/jsch-0.2.8)
   * [#287](https://github.com/mwiede/jsch/issues/287) add algorithm type information to algorithm negotiation logs.
   * [#289](https://github.com/mwiede/jsch/issues/289) wrap NoClassDefFoundError's for invalid private keys.